Refactor: 별점 계산 함수를 파일 분리
